pyproject.toml
[project]
name = "azure-foundry-ai-evaluator-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "azure-ai-evaluation==1.13.1",
    "azure-ai-projects==2.0.0b1",
    "lagom>=2.7.7,<3.0",
    "pydantic-settings>=2.11.0,<3.0",
]

[dependency-groups]
cli = [
    "pyright>=1.1.407,<2.0",
    "ruff>=0.14.4,<0.15.0",
]
dev = [
    "pyhamcrest>=2.1.0,<3.0",
    "pytest-cov>=7.0.0,<8.0",
    "pytest-dotenv>=0.5.2,<0.6.0",
    "pytest>=9.0.2,<10.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/ai_evaluator"
]
